About: extract page into more atomic components
