+++ components/About/About.tsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import AboutUniversiteams from './AboutUniversiteams'
+import AboutTeam120 from './AboutTeam120'
+
+const About = () => {
+  return (
+    <>
+      <AboutUniversiteams />
+      <AboutTeam120 />
+    </>
+  )
+}
+
+export default About
